% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rpart.R
\name{run_rpart}
\alias{run_rpart}
\title{Run rpart on a named gene, return clinical data with appended high/low based on cutoff}
\usage{
run_rpart(
  expr_df,
  gene_ids,
  clin_df,
  surv_event,
  surv_time,
  join_el = NULL,
  print_pdf = NULL,
  title_text = NULL
)
}
\arguments{
\item{expr_df}{data.frame containing expression in log2tpm with a unique column represented by param gene_id, rows are samples}

\item{gene_ids}{vector of strings in colnames(expr_df) relating to log2tpm data for one or more gene}

\item{clin_df}{data.frame containing clinical data with unique columns represented by surv_event and surv_time, rows are samples}

\item{surv_event}{colnames(clin_df) relating to survival event}

\item{surv_time}{colnames(clin_df) relating to survival event}

\item{join_el}{colname on which to join expression and survival data (default: rownames)}

\item{print_pdf}{print PDF to file (else return in output list)}

\item{title_text}{title text for plot}
}
\value{
list containing: [\link{1}] tibble of clin_df with columns appended: 'gene_id'_group, 'gene_id'_log2tpm; [\link{2}] named list of rpart plots by gene_ids
}
\description{
Run rpart on a named gene, return clinical data with appended high/low based on cutoff
}
\examples{

expr_df <- readRDS(system.file("extdata", "expr_df.rds", package="rpartSurvivalClassifier"))
clin_df <- readRDS(system.file("extdata", "clin_df.rds", package="rpartSurvivalClassifier"))
clin_new_tb <- rpartSurvivalClassifier::run_rpart(expr_df = expr_df, gene_id = "CRABP2", clin_df = clin_df, surv_event = "OS", surv_time = "OS.time", join_el = "sample")

}
