% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_surv_plot.R
\name{run_surv_plot}
\alias{run_surv_plot}
\title{Run survival analysis and plot using rpart output from run_rpart()}
\usage{
run_surv_plot(clin_tb, gene_id, surv_event, surv_time)
}
\arguments{
\item{clin_tb}{tibbe created by run_rpart()}

\item{gene_id}{string in colnames(expr_df) relating to log2tpm data}

\item{surv_event}{colnames(clin_df) relating to survival event}

\item{surv_time}{colnames(clin_df) relating to survival event}
}
\value{
tibble of clin_df with two columns appended, 'gene_id'_group, 'gene_id'_log2tpm
}
\description{
Run survival analysis and plot using rpart output from run_rpart()
}
\examples{

expr_df <- readRDS(system.file("extdata", "expr_df.rds", package="rpartSurvivalClassifier"))
clin_df <- readRDS(system.file("extdata", "clin_df.rds", package="rpartSurvivalClassifier"))
clin_new_tb <- run_rpart(expr_df = expr_df, gene_id = "CRABP2", clin_df = clin_df, surv_event = "OS", surv_time = "OS.time", join_el = "sample")
log_rank_res <- run_surv_plot(clin_tb = clin_new_tb, gene_id = "CRABP2", surv_event = "OS", surv_time = "OS.time")

}
