% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rpart.R
\name{run_rpart}
\alias{run_rpart}
\title{Run rpart on a named gene, return clinical data with appended high/low based on cutoff}
\usage{
run_rpart(expr_df, gene_id, clin_df, surv_event, surv_time, join_el = NULL)
}
\arguments{
\item{expr_df}{data.frame containing expression in log2tpm with a unique column represented by param gene_id, rows are samples}

\item{gene_id}{string in colnames(expr_df) relating to log2tpm data}

\item{clin_df}{data.frame containing clinical data with unique columns represented by surv_event and surv_time, rows are samples}

\item{surv_event}{colnames(clin_df) relating to survival event}

\item{surv_time}{colnames(clin_df) relating to survival event}

\item{join_el}{colname on which to join expression and survival data (default: rownames)}
}
\value{
tibble of clin_df with two columns appended, 'gene_id'_group, 'gene_id'_log2tpm
}
\description{
Run rpart on a named gene, return clinical data with appended high/low based on cutoff
}
\examples{

expr_df <- readRDS(system.file("extdata", "expr_df.rds", package="rpartSurvivalClassifier"))
clin_df <- readRDS(system.file("extdata", "clin_df.rds", package="rpartSurvivalClassifier"))
clin_new_tb <- run_rpart(expr_df = expr_df, gene_id = "CRABP2", clin_df = clin_df, surv_event = "OS", surv_time = "OS.time", join_el = "sample")

}
