% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_surv_plot.R
\name{run_surv_plot}
\alias{run_surv_plot}
\title{Run survival analysis and plot using rpart output from run_rpart()}
\usage{
run_surv_plot(
  clin_tb,
  gene_ids,
  surv_event,
  surv_time,
  expr_unit = "log2tpm",
  group_name = "_group",
  col_palette = NULL,
  print_pdf = NULL,
  print_png = NULL,
  title_text = "",
  sub_text = "",
  plot_prefix = NULL
)
}
\arguments{
\item{clin_tb}{tibble created by run_rpart(), or a tibble containing a columns named '{gene_ids\link{1}}_{group_name}' and {surv_event, surv_time} as below}

\item{gene_ids}{vector of strings used in run_rpart to define gene used in classification; if named vector, uses names and ids for plotting}

\item{surv_event}{colnames(clin_tb) relating to survival event}

\item{surv_time}{colnames(clin_tb) relating to survival event}

\item{expr_unit}{unit of expression in clin_tb; default - log2tpm}

\item{group_name}{character string of the 'group_name' column, therefore the gene_id concatenated with _{group_name} suffix; default: group}

\item{col_palette}{colours to use in plotting (vector, high -> low expression; think palette in ggsurvplot is alphanum sorted...)}

\item{print_pdf}{print PDF to file (else return in output list)}

\item{print_png}{print PDF to file (else return in output list)}

\item{title_text}{title text for plot}

\item{sub_text}{sub text for plot}

\item{plot_prefix}{prefix for plot filename}
}
\value{
table from survival::survdiff (log rank test), ggsurvplot PDF printed
}
\description{
Run survival analysis and plot using rpart output from run_rpart()
}
\examples{

expr_df <- readRDS(system.file("extdata", "expr_df.rds", package="rpartSurvivalClassifier"))
clin_df <- readRDS(system.file("extdata", "clin_df.rds", package="rpartSurvivalClassifier"))
clin_new_tb <- rpartSurvivalClassifier::run_rpart(expr_df = expr_df, gene_id = "CRABP2", clin_df = clin_df, surv_event = "OS", surv_time = "OS.time", join_el = "sample")
log_rank_res <- rpartSurvivalClassifier::run_surv_plot(clin_tb = clin_new_tb, gene_id = "CRABP2", surv_event = "OS", surv_time = "OS.time")

}
